requirements.txt
asgiref==3.8.1
Django==5.0.4
django-jazzmin==2.6.1
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
PyJWT==2.8.0
shortuuid==1.0.13
sqlparse==0.4.4
